Implement optional release_datadir input
...for BiobankScrubWorkflow.

This optional input designates the root of the directory tree where
the scrubbed files will be pushed.  It defaults to the value specified
for the current_datadir input.

The primary use-case for specifying a value for this input occurs
during testing, when it is often convenient to leave the original
current_datadir tree unchanged across repeated testing runs.