Make some scrub tasks ignore exit status
Specifically, add the

    ignoreExitStatus: true

directive to the `runtime` block of the ScrubBatch, ConcatenateShards,
and PushScrubbed tasks.

The rationale for this is to keep a run going when a fraction of the
calls in a scatter fails.  (If *all* the calls in a scatter fail, then
letting the rest of the run finish will be useless, but also
harmless.)