Adjust ConcatenateShards' "preemptible" setting
Of all of the pipeline's tasks, ConcatenateShards is the one that
performs the longest-running operations.  This means that this is the
task for which preemption results in the greatest delays.  Therefore,
in this commit we reduce the value of the preemptible parameter for
this task from 10 to 3.  (With this setting, the most recent run of
this workflow would have ultimately used non-preemptible nodes in 4
out of 67 calls of this task.)