Init commit, created joint calling feature
Added CTBHaplotypeCallerGvcf.wdl which runs the first few
steps of the BGWGS pipline but creates a gvcf instead of
a vcf file as output.

Added JointCalling.wdl which will run joint calling of
the gvcf files. It will run the GenerateSampleMap wdl
to create a sample map of all files in the input batch.
Then run the JointGenotyping wdl to jointly call all
input gvcf files and produce a jointly called output
vcf.