Fix bug in RunPRS.GetRegions
Due to a previous change, the outputs of the ExtractQueryVariants and
ExtractReferenceVariants tasks are now variant IDs that *always* begin
with a "bare" chromosome designation (e.g. 12 or X), irrespective of
whether or not the original VCF used chromosome labels featuring a
"chr" prefix.

In order to be consistent with this convention, the version of the
code committed here strips off any "chr" prefixes from the variant IDs
in the PCA variants and weights files.