Implement required MakeAdjustmentModel query input
The query_vcf and query_variants inputs of the workflow's earlier
version were both optional, though this is misleading, since exactly
one of them had to be specified.

The motivation for the change in this commit is to replace the two
optional arguments query_vcf and query_variants with a single
mandatory one, query_file.  The WDL will then handle query_file as a
VCF if its basename ends in ".vcf.gz"; otherwise, it will treat it as
a text file.