Have HelperTasks.Rename* tasks preserve basenames
Before these tasks obtained the basename of their output files by
prepending the prefix "renamed_" to the basenames of the input files.
In their new version, these tasks give to their output files their
input files' basenames.  (These output files are always in a director
separate from the input files, so collisions are not possible.)

The rationale for this change is that some downstream WDLs rely on the
information in some of these basenames for their operations.