Set batch size to 1
Added the --batch 1 flag to the lmm_variant_detection_report.py
call, this will make it so that instead of loading the default
50 variant per batch when pulling GI for variant annotations
the program will querry a signle variant at a time. This should
increase runtime significantly but allow the granualr level of
debugging of the server side logs to isolate the problamatic
variant producing the this position maps to multiple variants
erros state.