Use "NA" in mix score dosage and avg columns
The "NAMED_ALLELE_DOSAGE_SUM" column and "SCORE1_AVG"
column are not necessary for mix score and future
analysis with scores. However, the rest of the
pipeline expects the format of the .sscore file to
have these columns. Therefore, when calculating the
mix score sum, use "NA" as placeholders in these two
unnecessary columns.