Updated CreateFileBatches
CreateFileBatches was not corretly batching
due to design and arathemic bugs, the new
code retools the batching. Added the size_bytes
field to the JSON output to ensure this
information is available for the next task.
Added gigabyte conversion to the batch summary
output to make it easier to verify batch sizes.
Eliminated operations that would truncate
floating-point values.